move WSClient to it's own class
